chore(frontend): organize components
